requirements.txt
discord.py==0.16.12
mecab-python30.996.1
